+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.0.5] - 2026-09-07
+
+### Changed
+- Constants, SO names
+### Added
+- Copy Config file and test data menu items.
+
 ## [0.0.4] - 2026-09-07
 
 ### Added
@@ -13,7 +20,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - If currency icon is not present, use currency name text.
 ### Fixed
 - UI fonts/buttons sizes.
-
 
 ## [0.0.3] - 2026-09-06